Added new idea of: format_checks - for common/general checks that happen before template comparison
