+++ checksit/etc/checksit.ini
@@ -22,6 +22,7 @@ excludes =
 #global_attributes:creation_date global_attributes:version
 default_template_cache_dir = template-cache
 template_cache_dirs = template-cache other-cache
+default_template = template-cache/ceda-base.yml
 
 [dataset:ukmo-assim]
 regex_file = ukmo-nwp-strat_gbl-std_\d{10}_u-v-gph-t-w\.pp
