+++ checksit/format_checks/__init__.py
@@ -1,0 +1,5 @@
+from .format_checks import *
+
+from ..config import get_config
+conf = get_config()
+
