+++ checksit/format_checks/format_checks.py
@@ -1,0 +1,6 @@
+from ..config import get_config
+conf = get_config()
+
+
+class FormatChecker:
+    pass
