+++ cli_tests.txt
@@ -9,9 +9,7 @@ checksit check --template=/badc/ukcp18/data/land-cpm/uk/2.2km/rcp85/08/rss/day/l
 # Switch on automatic-caching when checking a file
 checksit check --auto-cache --template=/badc/ukcp18/data/land-cpm/uk/2.2km/rcp85/08/rss/day/latest/rss_rcp85_land-cpm_uk_2.2km_08_day_20671201-20681130.nc /badc/ukcp18/data/land-cpm/uk/2.2km/rcp85/01/rss/day/latest/rss_rcp85_land-cpm_uk_2.2km_01_day_20671201-20681130.nc
 
-checksit check --verbose /group_workspaces/jasmin2/ukcp18/UKcordex-laura/tasmax_rcp85_land-rcm_uk_12km_EC-EARTH_r12i1p1_HIRHAM5_day_19801201-19901130.nc
-
-checksit check /badc/ukcp09/data/gridded-land-obs/gridded-land-obs-monthly/grid/ascii/rainfall/2016/ukcp09_gridded-land-obs-monthly_5km_rainfall_201610.txt
+checksit check --verbose /group_workspaces/jasmin2/ukcp18/incoming-astephen/ukcordex-example/tasmax_rcp85_land-rcm_uk_12km_EC-EARTH_r12i1p1_HIRHAM5_day_19801201-19901130.nc
 
 checksit check /badc/ukmo-assim/data/standard/2022/ukmo-nwp-strat_gbl-std_2022020112_u-v-gph-t-w.pp
 
@@ -41,3 +39,6 @@ checksit check --rules=global_attributes:project=rule-func:match-one-of:"ukcp18|
 
 # Match string length rules
 checksit check --rules=global_attributes:project=rule-func:string-of-length:10,global_attributes:contact=rule-func:string-of-length:100+ /gws/nopw/j04/cmip6_prep_vol1/ukcp18/data/land-prob/v20211110/uk/25km/rcp85/sample/b8110/30y/cltAnom/mon/v20211110/cltAnom_rcp85_land-prob_uk_25km_sample_b8110_30y_mon_20091201-20991130.nc 
+
+# The config can include a default template, if none are found, here it uses: template-cache/ceda-base.yml
+checksit check  /home/users/astephen/summer_rainfall_2001.nc
