+++ run-cli-tests.sh
@@ -1,9 +1,34 @@
 #!/bin/bash
 
-while read CMD ; do
+prompt=false
+[ "$1" == "--prompt" ] || [ "$1" == "-p" ] && prompt=true
+
+mapfile -t lines < cli_tests.txt
+
+for i in $(seq ${#lines[@]}) ; do 
+    CMD=${lines[$i]} 
+
     echo
     [ ! "$CMD" ] || [[ "$CMD" =~ \# ]] && continue
-    echo "[INFO] Running: $CMD"
+    echo "[INFO] Running: "
+    
+    count=0
+    for i in $(echo $CMD); do 
+        [ $count -eq 0 ] && echo "$ checksit check"
+        [ $count -ge 2 ] && echo "     $i"
+        let count+=1
+    done
+
+    if [ $prompt == "true" ]; then
+        echo
+        echo "Press ENTER to continue..."
+        read waiter
+        if [ "$waiter" == "q" ]; then
+            echo "[INFO] Exiting!"
+            exit
+        fi
+    fi
+ 
     $CMD
 
     if [ $? -ne 0 ]; then
@@ -11,5 +36,8 @@ while read CMD ; do
         echo "[ERROR] The following command failed...so aborting: $CMD"
         break
     fi
-done < cli_tests.txt
+
+    echo
+    echo "=========================================================================="
+done
